changed client addObjective implementation to be idempotent
