+++ client/src/SessionView/index.js
@@ -1,4 +1,4 @@
-import { useEffect, useContext, useState, useCallback } from 'react'
+import { useState, useCallback } from 'react'
 import {
   Avatar,
   Card,
@@ -14,9 +14,8 @@ import { LocalLibrary, PhotoLibrary } from '@material-ui/icons'
 import { makeStyles } from '@material-ui/core/styles'
 
 import * as factions from '../gameInfo/factions'
-import { SignalRConnectionContext } from '../signalR'
-import { DispatchContext } from '../state'
 
+import useRealTimeSession from './useRealTimeSession'
 import ShuffleFactionsButton from './ShuffleFactionsButton'
 import ShareButton from './ShareButton'
 import VictoryPoints from './VictoryPoints'
@@ -39,29 +38,6 @@ const useStyles = makeStyles(theme => ({
     cursor: 'pointer',
   },
 }))
-
-const useRealTimeSession = sessionId => {
-  const signalRConnection = useContext(SignalRConnectionContext)
-  const dispatch = useContext(DispatchContext)
-
-  useEffect(() => {
-    signalRConnection.on("SessionEvent", sessionEvent => {
-      const payload = JSON.parse(sessionEvent.serializedPayload)
-      dispatch({ type: sessionEvent.eventType, payload })
-    })
-  }, [signalRConnection, dispatch])
-
-  useEffect(() => {
-    if (!sessionId) {
-      return
-    }
-
-    signalRConnection.invoke("UnsubscribeFromSession", sessionId)
-    signalRConnection.invoke("SubscribeToSession", sessionId)
-
-    return () => signalRConnection.invoke("UnsubscribeFromSession", sessionId)
-  }, [signalRConnection, sessionId])
-}
 
 const getFactionCheatSheetPath = factionKey => `/factionCheatsheets/${factionKey.toLowerCase()}.png`
